+++ inst/testing/testing.R
@@ -2,7 +2,6 @@
 # Prerequisites -----------------------------------------------------------
 
 library(dplyr)
-library(stringr)
 library(devtools)
 load_all()
 
@@ -14,5 +13,4 @@ load_all()
 # TODO Add another function to evaluate assays, including their test_cutoff
 # TODO Should ab_titer_unit remain an optional field?
 
-# TODO Update region's data with Jane's work